backend/requirements.txt
# FastAPI and server dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data validation and serialization
pydantic[email]>=2.5.0
python-dotenv>=1.0.0

# Database and ORM
sqlalchemy>=2.0.0
alembic>=1.12.0

# HTTP and networking
requests>=2.31.0
aiohttp>=3.9.0
aiofiles>=23.2.0

# Image processing
pillow>=10.0.0

# Date and time utilities
python-dateutil>=2.8.0

# AI and LangChain dependencies
langgraph>=0.2.55
langchain>=0.3.7
langchain-openai>=0.2.10
langchain-community>=0.3.5

# Observability and tracing (optional)
arize-otel>=0.8.1
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp>=1.21.0
openinference-instrumentation-openai>=0.1.0
openinference-instrumentation-langchain>=0.1.0 